requirements.txt
requests
flatten_json
openai
python-dotenv
clickhouse-driver
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-multipart
httpx
tiktoken
tqdm